feat(release): Release should capture ALL app values.
When an app is added to a release, all the app's values (including values in files referenced from
the app config) are copied into the app release config. This makes the release self-contained - it
does not rely on any files from the app repo, and accurately preserves what is actually released
when the release is deployed.